N was being squared as an int before being cast into a long, causing overflow in some cases. Fixed.

+++ src/unl/cse/kaprekar/KaprekarUtils.java
@@ -29,7 +29,7 @@ public class KaprekarUtils {
 		} 
 		
 		//we use a long to accommodate larger squares
-		long square = n * n;
+		long square = n * (long)n;
 		int numDigits = (int) Math.log10(square) + 1;
 		long modulus = 1;
 		long first, second;
